More actionFinalize
